% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_ranking_data}
\alias{raw_ranking_data}
\title{raw_ranking_data}
\format{
A data frame with 34 rows and 166 variables:
\describe{
\item{\code{asf}}{asf}
}
}
\source{
\url{https://www.sciencedirect.com/science/article/pii/S1048984318303746?via\%3Dihub}
}
\usage{
raw_ranking_data
}
\description{
Peer-rated Data on Psycho-Social Traits From a Chabu Community
}
\keyword{datasets}
